+++ src/main/java/org/markdownwriterfx/preview/HtmlSourcePreview.java
@@ -6,11 +6,11 @@
  * modification, are permitted provided that the following conditions are met:
  *
  *  o Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
+ *	notice, this list of conditions and the following disclaimer.
  *
  *  o Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
+ *	notice, this list of conditions and the following disclaimer in the
+ *	documentation and/or other materials provided with the distribution.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -28,8 +28,15 @@
 package org.markdownwriterfx.preview;
 
 import java.nio.file.Path;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import javafx.scene.control.ScrollBar;
-import javafx.scene.control.TextArea;
+import org.fxmisc.flowless.VirtualizedScrollPane;
+import org.fxmisc.richtext.StyleClassedTextArea;
+import org.fxmisc.richtext.model.StyleSpans;
+import org.fxmisc.richtext.model.StyleSpansBuilder;
 import org.markdownwriterfx.preview.MarkdownPreviewPane.Renderer;
 import org.markdownwriterfx.util.Utils;
 
@@ -41,37 +48,103 @@ import org.markdownwriterfx.util.Utils;
 class HtmlSourcePreview
 	implements MarkdownPreviewPane.Preview
 {
-	private final TextArea textArea = new TextArea();
+	private final StyleClassedTextArea textArea = new StyleClassedTextArea();
+	private final VirtualizedScrollPane<StyleClassedTextArea> scrollPane = new VirtualizedScrollPane<>(textArea);
 	private ScrollBar vScrollBar;
 
 	HtmlSourcePreview() {
 		textArea.setEditable(false);
 		textArea.setFocusTraversable(false);
 		textArea.setWrapText(true);
+		textArea.getStylesheets().add("org/markdownwriterfx/preview/HtmlSourcePreview.css");
 	}
 
 	@Override
 	public javafx.scene.Node getNode() {
-		return textArea;
+		return scrollPane;
 	}
 
 	@Override
 	public void update(Renderer renderer, Path path) {
-		double scrollTop = textArea.getScrollTop();
+		double scrollY = textArea.getEstimatedScrollY();
 
-		textArea.setText(renderer.getHtml());
+		String html = renderer.getHtml();
+		textArea.replaceText(html);
+		textArea.setStyleSpans(0, computeHighlighting(html));
 
-		textArea.setScrollTop(scrollTop);
+		textArea.setEstimatedScrollY(scrollY);
 	}
 
 	@Override
 	public void scrollY(double value) {
 		if (vScrollBar == null)
-			vScrollBar = Utils.findVScrollBar(textArea);
+			vScrollBar = Utils.findVScrollBar(scrollPane);
 		if (vScrollBar == null)
 			return;
 
 		double maxValue = vScrollBar.maxProperty().get();
 		vScrollBar.setValue(maxValue * value);
+	}
+
+	//---- XML syntax highlighter ---------------------------------------------
+
+	// from richtextfx-demos/src/main/java/org/fxmisc/richtext/demo/XMLEditor.java
+
+	private static final Pattern XML_TAG = Pattern.compile("(?<ELEMENT>(</?\\h*)(\\w+)([^<>]*)(\\h*/?>))"
+			+"|(?<COMMENT><!--[^<>]+-->)");
+
+	private static final Pattern ATTRIBUTES = Pattern.compile("(\\w+\\h*)(=)(\\h*\"[^\"]+\")");
+
+	private static final int GROUP_OPEN_BRACKET = 2;
+	private static final int GROUP_ELEMENT_NAME = 3;
+	private static final int GROUP_ATTRIBUTES_SECTION = 4;
+	private static final int GROUP_CLOSE_BRACKET = 5;
+	private static final int GROUP_ATTRIBUTE_NAME = 1;
+	private static final int GROUP_EQUAL_SYMBOL = 2;
+	private static final int GROUP_ATTRIBUTE_VALUE = 3;
+
+	private static StyleSpans<Collection<String>> computeHighlighting(String text) {
+
+		Matcher matcher = XML_TAG.matcher(text);
+		int lastKwEnd = 0;
+		StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
+		while(matcher.find()) {
+
+			spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
+			if(matcher.group("COMMENT") != null) {
+				spansBuilder.add(Collections.singleton("comment"), matcher.end() - matcher.start());
+			}
+			else {
+				if(matcher.group("ELEMENT") != null) {
+					String attributesText = matcher.group(GROUP_ATTRIBUTES_SECTION);
+
+					spansBuilder.add(Collections.singleton("tagmark"), matcher.end(GROUP_OPEN_BRACKET) - matcher.start(GROUP_OPEN_BRACKET));
+					spansBuilder.add(Collections.singleton("anytag"), matcher.end(GROUP_ELEMENT_NAME) - matcher.end(GROUP_OPEN_BRACKET));
+
+					if(!attributesText.isEmpty()) {
+
+						lastKwEnd = 0;
+
+						Matcher amatcher = ATTRIBUTES.matcher(attributesText);
+						while(amatcher.find()) {
+							spansBuilder.add(Collections.emptyList(), amatcher.start() - lastKwEnd);
+							spansBuilder.add(Collections.singleton("attribute"), amatcher.end(GROUP_ATTRIBUTE_NAME) - amatcher.start(GROUP_ATTRIBUTE_NAME));
+							spansBuilder.add(Collections.singleton("tagmark"), amatcher.end(GROUP_EQUAL_SYMBOL) - amatcher.end(GROUP_ATTRIBUTE_NAME));
+							spansBuilder.add(Collections.singleton("avalue"), amatcher.end(GROUP_ATTRIBUTE_VALUE) - amatcher.end(GROUP_EQUAL_SYMBOL));
+							lastKwEnd = amatcher.end();
+						}
+						if(attributesText.length() > lastKwEnd)
+							spansBuilder.add(Collections.emptyList(), attributesText.length() - lastKwEnd);
+					}
+
+					lastKwEnd = matcher.end(GROUP_ATTRIBUTES_SECTION);
+
+					spansBuilder.add(Collections.singleton("tagmark"), matcher.end(GROUP_CLOSE_BRACKET) - lastKwEnd);
+				}
+			}
+			lastKwEnd = matcher.end();
+		}
+		spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
+		return spansBuilder.create();
 	}
 }